Refactors verify
